requirements-dev.txt

codecov==2.0.9
coverage==4.5.2
flake8==3.7.9
mock==2.0.0
pytest>=3.9.1
pytest-cov>=2.6.0
pytest-django
